+++ proto/session/encoding.go
@@ -46,7 +46,7 @@ const (
 func (x *ObjectSessionContext) MarshaledSize() int {
 	var sz int
 	if x != nil {
-		sz = proto.SizeVarint(FieldObjectSessionContextVerb, int32(x.Verb)) +
+		sz = proto.SizeVarint(FieldObjectSessionContextVerb, x.Verb) +
 			proto.SizeEmbedded(FieldObjectSessionContextTarget, x.Target)
 	}
 	return sz
@@ -58,7 +58,7 @@ func (x *ObjectSessionContext) MarshaledSize() int {
 // NPE-safe.
 func (x *ObjectSessionContext) MarshalStable(b []byte) {
 	if x != nil {
-		off := proto.MarshalToVarint(b, FieldObjectSessionContextVerb, int32(x.Verb))
+		off := proto.MarshalToVarint(b, FieldObjectSessionContextVerb, x.Verb)
 		proto.MarshalToEmbedded(b[off:], FieldObjectSessionContextTarget, x.Target)
 	}
 }
@@ -76,7 +76,7 @@ const (
 func (x *ContainerSessionContext) MarshaledSize() int {
 	var sz int
 	if x != nil {
-		sz = proto.SizeVarint(FieldContainerSessionContextVerb, int32(x.Verb)) +
+		sz = proto.SizeVarint(FieldContainerSessionContextVerb, x.Verb) +
 			proto.SizeBool(FieldContainerSessionContextWildcard, x.Wildcard) +
 			proto.SizeEmbedded(FieldContainerSessionContextContainerID, x.ContainerId)
 	}
@@ -89,7 +89,7 @@ func (x *ContainerSessionContext) MarshaledSize() int {
 // MarshalStable is NPE-safe.
 func (x *ContainerSessionContext) MarshalStable(b []byte) {
 	if x != nil {
-		off := proto.MarshalToVarint(b, FieldContainerSessionContextVerb, int32(x.Verb))
+		off := proto.MarshalToVarint(b, FieldContainerSessionContextVerb, x.Verb)
 		off += proto.MarshalToBool(b[off:], FieldContainerSessionContextWildcard, x.Wildcard)
 		proto.MarshalToEmbedded(b[off:], FieldContainerSessionContextContainerID, x.ContainerId)
 	}
